+++ src/app/dashboard/services/product-list/product-list.component.ts
@@ -3,6 +3,9 @@ import { ActivatedRoute } from '@angular/router';
 import { ApiService } from 'src/app/shared/services/api.service';
 import tableData from '../../../../assets/data/pages/table-data.json';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { NzModalService } from 'ng-zorro-antd/modal';
+import { NotificationService } from 'src/app/shared/services/notification.service';
+import { environment } from 'src/environments/environment';
 
 @Component({
   selector: 'app-product-list',
@@ -11,7 +14,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 })
 export class ProductListComponent implements OnInit {
   serviceID;
-  host = 'http://localhost:8080';
+  host = environment.base_URL;
   isEditting: boolean = false;
   seller: any;
   productData = [];
@@ -45,7 +48,9 @@ export class ProductListComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private apiService: ApiService,
-    private fb: FormBuilder
+    private fb: FormBuilder,
+    private notificationService: NotificationService,
+    private modalService: NzModalService
   ) {
     this.serviceID = this.route.snapshot.params['Serviceid'];
     this.seller = tableData.bestSeller;
@@ -111,18 +116,38 @@ export class ProductListComponent implements OnInit {
     if (this.productForm.valid) {
       let id = this.productForm.get('id').value;
       let productData = this.productForm.value;
-      if (id) {
-        this.apiService.updateProduct(productData).subscribe((_) => {
-          this.isOpenProductModal = false;
-          this.fetchProductData();
-        });
-        return;
-      }
-      this.apiService.createProduct(productData).subscribe((_) => {
-        this.isOpenProductModal = false;
-        this.fetchProductData();
+      this.modalService.confirm({
+        nzTitle: `Do you Want to ${
+          this.isEditting ? 'Update' : 'Create'
+        } this Product?`,
+        nzContent: `Product Name: ${this.productForm.get('name').value}`,
+        nzOnOk: () => {
+          if (id) {
+            this.apiService.updateProduct(productData).subscribe((_) => {
+              this.isOpenProductModal = false;
+              this.fetchProductData();
+              this.notificationService.createNotification({
+                type: 'success',
+                title: 'Product Updated',
+                message: `Product ${productData.name} has been updated`,
+                position: 'bottomRight',
+              })
+            });
+            return;
+          }
+          this.apiService.createProduct(productData).subscribe((_) => {
+            this.isOpenProductModal = false;
+            this.fetchProductData();
+            this.notificationService.createNotification({
+              type: 'success',
+              title: 'Product Created',
+              message: `Product ${productData.name} has been created`,
+              position: 'bottomRight',
+            })
+          });
+          return;
+        },
       });
-      return;
     }
   }
   onUploadImage(event) {
@@ -136,9 +161,22 @@ export class ProductListComponent implements OnInit {
     }
   }
   onDeleteProduct(id) {
-    this.apiService.deleteProduct(id).subscribe((_) => {
-      this.fetchProductData();
-    });
+    this.modalService.confirm({
+      nzTitle: `Do you Want to Delete this Product?`,
+      nzContent: `Product ID: ${this.productForm.get('name').value}`,
+      nzOnOk: () => {
+        this.apiService.deleteProduct(id).subscribe((_) => {
+          this.fetchProductData();
+          this.notificationService.createNotification({
+            type: 'error',
+            title: 'Product Deleted',
+            message: `Product has been deleted`,
+            position: 'bottomRight',
+          })
+        });
+      },
+    })
+
   }
   filterProductData(e) {
     if (e == '') {
